# stock-python/contracts/openapi/public_api.yaml
openapi: 3.0.3
info:
  title: Stock API - Public
  description: Public API endpoints for stock trading platform
  version: 1.0.0
  contact:
    name: Stock API Support
servers:
  - url: /api/v1
    description: Public API v1
...
paths:
...
  /signals:
    get:
      summary: List signals
      tags: [signals]
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: signal_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of signals
    post:
      summary: Create signal manually
      tags: [signals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stock_id:
                  type: integer
                signal_type:
                  type: string
                entry_price:
                  type: number
                stop_loss:
                  type: number
                take_profit_1:
                  type: number
                take_profit_2:
                  type: number
                take_profit_3:
                  type: number
                reasoning:
                  type: string
      responses:
        '200':
          description: Signal created
  /signals/active:
    get:
      summary: Get active signals
      tags: [signals]
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of active signals
...
  /signals/generate:
    post:
      summary: Generate signal from OHLCV data
      tags: [signals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symbol:
                  type: string
                high:
                  type: array
                  items:
                    type: number
                low:
                  type: array
                  items:
                    type: number
                close:
                  type: array
                  items:
                    type: number
                volume:
                  type: array
                  items:
                    type: integer
                strategy_id:
                  type: integer
      responses:
        '200':
          description: Signal generated
  /signals/{signal_id}:
    patch:
      summary: Update signal
      tags: [signals]
      parameters:
        - name: signal_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                stop_loss:
                  type: number
                take_profit_1:
                  type: number
                take_profit_2:
                  type: number
                take_profit_3:
                  type: number
      responses:
        '200':
          description: Signal updated
  /signals/{signal_id}/trigger:
    post:
      summary: Trigger signal
      tags: [signals]
      parameters:
        - name: signal_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Signal triggered
  /signals/{signal_id}/expire:
    post:
      summary: Expire signal
      tags: [signals]
      parameters:
        - name: signal_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Signal expired
  /signals/{signal_id}/cancel:
    post:
      summary: Cancel signal
      tags: [signals]
      parameters:
        - name: signal_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Signal cancelled
  /signals/history/range:
    get:
      summary: Get signals by date range
      tags: [signals]
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
        - name: end_date
          in: query
          required: true
          schema:
            type: string
        - name: signal_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Signals in date range

  # Notifications
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...